# swagger-schema.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/history/list": {
      "get": {
        "operationId": "HistoryController_findAllHistory",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 10,
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term for sessionId or observations",
            "schema": {
              "example": "session_123",
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter by user ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "description": "Filter by customer ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "name": "protocol",
            "required": false,
            "in": "query",
            "description": "Filter by protocol",
            "schema": {
              "example": "20251218175106819",
              "type": "string"
            }
          },
          {
            "name": "platform",
            "required": false,
            "in": "query",
            "description": "Filter by platform",
            "schema": {
              "example": "whatsapp",
              "type": "string",
              "enum": [
                "whatsapp",
                "telegram",
                "instagram",
                "facebook",
                "other"
              ]
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Filter by start date (ISO string)",
            "schema": {
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Filter by end date (ISO string)",
            "schema": {
              "example": "2024-01-31T23:59:59.999Z",
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active interactions",
            "schema": {
              "example": true,
              "type": "string"
            }
          },
          {
            "name": "isAttended",
            "required": false,
            "in": "query",
            "description": "Filter by attended interactions",
            "schema": {
              "example": true,
              "type": "string"
            }
          },
          {
            "name": "isFinished",
            "required": false,
            "in": "query",
            "description": "Filter by finished interactions",
            "schema": {
              "example": true,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "History records retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/HistoryResponseDto"
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    },
                    "totalPages": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all history records with pagination and filtering",
        "tags": [
          "History"
        ]
      }
    },
...
    "/history/statistics": {
      "get": {
        "operationId": "HistoryController_getHistoryStatistics",
        "parameters": [
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Filter by end date (ISO string)",
            "schema": {}
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Filter by start date (ISO string)",
            "schema": {}
          },
          {
            "name": "platform",
            "required": false,
            "in": "query",
            "description": "Filter by platform",
            "schema": {}
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "description": "Filter by customer ID",
            "schema": {}
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter by user ID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "History statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "number"
                    },
                    "active": {
                      "type": "number"
                    },
                    "attended": {
                      "type": "number"
                    },
                    "finished": {
                      "type": "number"
                    },
                    "averageDuration": {
                      "type": "number"
                    },
                    "averageAttendanceTime": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get history statistics",
        "tags": [
          "History"
        ]
      }
    },
...
    "/chat/history": {
      "get": {
        "description": "Retrieve the chat history for a specific session from Redis storage. Reactions are updated in real-time and stored in Redis for fast access.",
        "operationId": "ChatController_getChatHistory",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "query",
            "description": "Session ID to retrieve chat history for",
            "schema": {
              "example": "session_123456789",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of messages to retrieve",
            "schema": {
              "default": 50,
              "example": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessionId": {
                      "type": "string",
                      "example": "session_123456789"
                    },
                    "messages": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "description": "Message object with all message properties including reactions",
                        "properties": {
                          "reactions": {
                            "type": "array",
                            "description": "Array of reactions for this message",
                            "items": {
                              "type": "object",
                              "properties": {
                                "emoji": {
                                  "type": "string",
                                  "example": "❤️"
                                },
                                "reactorId": {
                                  "type": "string",
                                  "example": "customer-uuid"
                                },
                                "reactedAt": {
                                  "type": "string",
                                  "format": "date-time"
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "count": {
                      "type": "number",
                      "example": 25
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid session ID or query parameters"
          },
          "401": {
            "description": "User not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get chat history for a session",
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/statistics": {
      "get": {
        "description": "Retrieve statistics about messages for a specific session including counts from both Redis and PostgreSQL.",
        "operationId": "ChatController_getSessionStatistics",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "query",
            "description": "Session ID to get statistics for",
            "schema": {
              "example": "session_123456789",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessionId": {
                      "type": "string",
                      "example": "session_123456789"
                    },
                    "redisCount": {
                      "type": "number",
                      "example": 25
                    },
                    "postgresCount": {
                      "type": "number",
                      "example": 100
                    },
                    "lastMessageAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2024-01-01T10:00:00.000Z"
                    },
                    "firstMessageAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2024-01-01T09:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid session ID"
          },
          "401": {
            "description": "User not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get session statistics",
        "tags": [
          "Chat"
        ]
      }
    },
...
  },
  "info": {
    "title": "CCS Backend API",
    "description": "Multi platform CRM API for customer service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "Typebot-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Typebot API Key (Bearer token)"
      }
    },
...
  }
}